// packages/backend/src/database/types/imovel.ts
/* tslint:disable */


/**
 * AUTO-GENERATED FILE @ 2020-06-21 04:26:57 - DO NOT EDIT!
 *
 * This file was automatically generated by schemats v.3.0.3
 * $ schemats generate -c postgres://username:password@localhost/reloaded -t imovel -s public
 *
 */


    export type id = number;
    export type descricao = string | null;
    export type valor_proposta = number | null;
    export type categoria = string | null;
    export type num_quartos = number | null;
    export type situacao = number | null;


export interface Imovel {
    id: id;
    descricao: descricao;
    valor_proposta: valor_proposta;
    categoria: categoria;
    num_quartos: num_quartos;
    situacao: situacao;

}