// packages/backend/src/database/types/transacao.ts
/* tslint:disable */


/**
 * AUTO-GENERATED FILE @ 2020-06-21 04:28:44 - DO NOT EDIT!
 *
 * This file was automatically generated by schemats v.3.0.3
 * $ schemats generate -c postgres://username:password@localhost/reloaded -t transacao -s public
 *
 */


    export type id = number;
    export type valor_total = number | null;
    export type data = Date | null;
    export type fk_pessoa_id = number | null;
    export type fk_imovel_id = number | null;

export interface Transacao {
    id: id;
    valor_total: valor_total;
    data: data;
    fk_pessoa_id: fk_pessoa_id;
    fk_imovel_id: fk_imovel_id;

}