// packages/backend/src/database/types/pessoa_transacao.ts
/* tslint:disable */


/**
 * AUTO-GENERATED FILE @ 2020-06-21 04:28:28 - DO NOT EDIT!
 *
 * This file was automatically generated by schemats v.3.0.3
 * $ schemats generate -c postgres://username:password@localhost/reloaded -t pessoa_transacao -s public
 *
 */


    export type fk_pessoa_id = number | null;
    export type fk_transacao_id = number | null;
    export type tipo = string | null;
    export type valor = number | null;

export interface PessoaTransacao {
    fk_pessoa_id: fk_pessoa_id;
    fk_transacao_id: fk_transacao_id;
    tipo: tipo;
    valor: valor;

}